requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
pdfplumber==0.10.3
requests==2.31.0
langchain>=0.1.0
langgraph==0.0.26
langchain-core>=0.1.25
langchain-community>=0.0.10
langchain-openai>=0.0.5
python-dotenv==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.3.1
transformers==4.37.2